requirements.txt
streamlit
python-dotenv
azure-cognitiveservices-speech
openai
nbformat
python-pptx
ffmpeg-python
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib